+++ psutil/arch/osx/proc_utils.c
@@ -97,6 +97,12 @@ psutil_sysctl_procargs(pid_t pid, char *procargs, size_t *argmax) {
             psutil_oserror_ad("sysctl(KERN_PROCARGS2) -> EIO");
             return -1;
         }
+        if (errno == 0) {
+            // see: https://github.com/giampaolo/psutil/issues/2708
+            psutil_debug("sysctl(KERN_PROCARGS2) -> errno 0");
+            psutil_oserror_ad("sysctl(KERN_PROCARGS2) -> errno 0");
+            return 0;
+        }
         psutil_oserror_wsyscall("sysctl(KERN_PROCARGS2)");
         return -1;
     }